Added dockerfile and K8 config
