+++ upwork-devs/ashish-belwase/virusshare_poc/Dockerfile
@@ -1,7 +1,7 @@
-FROM python:3-alpine
+FROM python:3.6
 EXPOSE 5000
 WORKDIR /usr/src/app
 COPY requirements.txt .
 RUN pip install -r requirements.txt
 COPY . .
-CMD [ "python", "src" ]
+CMD [ "python", "main.py" ]
