+++ upwork-devs/ashish-belwase/virusshare_poc/deploy.yaml
@@ -1,0 +1,35 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: k8-test-data-service
+spec:
+  selector:
+    app: k8-test-data
+  ports:
+  - protocol: "TCP"
+    port: 9000
+    targetPort: 5000
+  type: LoadBalancer
+
+---
+
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: k8-test-data
+spec:
+  selector:
+    matchLabels:
+      app: k8-test-data
+  replicas: 2
+  template:
+    metadata:
+      labels:
+        app: k8-test-data
+    spec:
+      containers:
+      - name: k8-test-data
+        image: belwase/k8-test-data
+        imagePullPolicy: Never
+        ports:
+        - containerPort: 5000
