+++ upwork-devs/ashish-belwase/virusshare_poc/main.py
@@ -1,0 +1,5 @@
+from src import create_app
+
+if __name__ == "__main__":
+    app = create_app()
+    app.run(debug=True, host="0.0.0.0")
