+++ upwork-devs/ashish-belwase/virusshare_poc/src/scrapers/__init__.py
@@ -1,1 +1,1 @@
-from scrapers.virusshare import VSScraper
+from src.scrapers.virusshare import VSScraper
